requirements.txt
torch==2.3.1
transformers==4.41.2
gradio==4.37.2
numpy==1.23.5
scipy==1.11.3
black
pytest
scikit-learn
sentencepiece==0.2.0
PyMuPDF==1.24.5  
pytesseract==0.3.10
pillow==10.3.0
protobuf